fix(dagger): fix invalid Python package name
