+++ .dagger-ci/daggerci/main.py
@@ -1,0 +1,58 @@
+#!/usr/bin/python
+'''
+Python script to build and test Docker containers for coreboot and EDK2 compilation
+'''
+
+# Logging
+# https://docs.python.org/3/howto/logging.html
+# DEBUG, INFO, WARNING, ERROR, CRITICAL
+import logging
+
+import os
+import sys
+import anyio
+
+from lib.cli import cli
+from lib.orchestrator import Orchestrator
+
+
+async def main(args: list = None):
+    '''
+    The main function, duh
+    '''
+    args, parser = cli(args=args)
+
+    # Setup nice logging
+    logging.basicConfig(format='%(levelname)s: %(message)s',
+                        level=logging.DEBUG if args.verbose else logging.INFO)
+
+    # Figure out location of docker compose
+    current_dir = os.path.dirname(os.path.realpath(__file__))
+    repo_root_dir = current_dir
+    while os.path.basename(repo_root_dir) != 'firmware-action':
+        repo_root_dir = os.path.dirname(repo_root_dir)
+    docker_compose_path = os.path.join(repo_root_dir, 'docker', 'compose.yaml')
+
+    # Init the Orchestrator
+    my_orchestrator = Orchestrator(
+        docker_compose_path=docker_compose_path,
+        concurent=args.concurent,
+        publish=args.publish
+    )
+
+    # Perform builds, tests and publishing
+    results = await my_orchestrator.build_test_publish(dockerfiles_override=args.dockerfile)
+
+    # Pretty print results
+    results.print()
+    return results.return_code
+
+
+if __name__ == '__main__':
+    #####
+    # Mandatory guard
+    # Detailed explanation: https://stackoverflow.com/a/419185
+    #####
+
+    # LET'S DO IT !!!
+    sys.exit(anyio.run(main))
